Added advanced exercise.

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,6 @@
 monthly_income = input("Enter your monthly income: ")
 monthly_expenses = input("Enter your monthly expenses: ")
-monthly_savings = int(monthly_income) - int(monthly_expenses)
+monthly_savings = float(monthly_income) - float(monthly_expenses)
 pro_savings = (monthly_savings * 12) + (monthly_savings * 12 * 0.05)
 print("Your monthly savings are $" + str(monthly_savings))
 print("Projected savings after one year, with interest, is: $" + str(pro_savings))
-
